feat(like): like api & add like-comment-scrap field to meeting

+++ src/main/java/swe/second/team_matching_server/domain/like/repository/UserMeetingLikeRepository.java
@@ -19,7 +19,7 @@ public interface UserMeetingLikeRepository extends JpaRepository<UserMeetingLike
   @Query("SELECT ums.user FROM UserMeetingLike ums WHERE ums.meeting.id = :meetingId")
   List<User> findLikedUsersByMeetingId(@Param("meetingId") Long meetingId);
 
-  long countByMeetingId(Long meetingId);
+  int countByMeetingId(Long meetingId);
 
   boolean existsByUserIdAndMeetingId(String userId, Long meetingId);
 }
